[GREEN-56] rename tax to documents
