+++ src/localisation/messages.js
@@ -22,7 +22,7 @@ export default Object.freeze({
       search: 'Search',
       information: 'Information',
       rules: 'Rules',
-      tax: 'Tax',
+      documents: 'Documents',
       payment: 'Payment',
       checkInDetails: 'CheckIn Details',
     },
@@ -144,7 +144,6 @@ export default Object.freeze({
       search: 'Поиск',
       information: 'Информация',
       rules: 'Правила',
-      tax: 'Налоги',
       payment: 'Оплата',
       checkInDetails: 'Детали регистрации',
     },
@@ -249,7 +248,6 @@ export default Object.freeze({
       search: '搜索',
       information: '信息',
       rules: '规则',
-      tax: '税',
       payment: '支付',
       checkInDetails: '入住详情',
     },
@@ -354,7 +352,6 @@ export default Object.freeze({
       search: 'Traži',
       information: 'Informacija',
       rules: 'Pravila',
-      tax: 'Porez',
       payment: 'Plaćanje',
       checkInDetails: 'Detalji prijave',
     },
@@ -459,7 +456,6 @@ export default Object.freeze({
       search: 'Recherche',
       information: 'Information',
       rules: 'Règles',
-      tax: 'Impôt',
       payment: 'Paiement',
       checkInDetails: 'Détails de l\'enregistrement',
     },
@@ -564,7 +560,6 @@ export default Object.freeze({
       search: 'Suchen',
       information: 'Information',
       rules: 'Regeln',
-      tax: 'Steuer',
       payment: 'Zahlung',
       checkInDetails: 'Check-In-Details',
     },
@@ -669,7 +664,6 @@ export default Object.freeze({
       search: 'Ricerca',
       information: 'Informazione',
       rules: 'Regole',
-      tax: 'Imposta',
       payment: 'Pagamento',
       checkInDetails: 'Dettagli del check-in',
     },
@@ -774,7 +768,6 @@ export default Object.freeze({
       search: 'Szukaj',
       information: 'Informacja',
       rules: 'Zasady',
-      tax: 'Podatek',
       payment: 'Zapłata',
       checkInDetails: 'Szczegóły zameldowania',
     },
@@ -879,7 +872,6 @@ export default Object.freeze({
       search: 'Iskanje',
       information: 'Informacije',
       rules: 'Pravila',
-      tax: 'Davek',
       payment: 'Plačilo',
       checkInDetails: 'Podrobnosti o prijavi',
     },
@@ -984,7 +976,6 @@ export default Object.freeze({
       search: 'Buscar',
       information: 'Información',
       rules: 'Normas',
-      tax: 'Impuesto',
       payment: 'Pago',
       checkInDetails: 'Detalles de registro',
     },
