--- src/views/Confirmation/Tax.vue
@@ -1,7 +1,0 @@
-<script setup>
-  import Tax from '@/components/Tax/Tax.vue';
-</script>
-
-<template>
-  <Tax />
-</template>
